src_ext/requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.0
flower==2.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.7.0
idna==3.4
kombu==5.3.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pydantic==2.0.2
pydantic-core==2.1.2
pydantic-settings==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tornado==6.3.2
typing-extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
